configuration: Move some more complex common types to their own files.
