+++ crates/configuration/src/lib.rs
@@ -9,5 +9,5 @@ pub use configuration::{
     as_runtime_configuration, configure, set_connection_uri, validate_raw_configuration,
     Configuration, RawConfiguration, RuntimeConfiguration,
 };
-pub use values::{ConnectionUri, ResolvedSecret};
-pub use version2::{occurring_scalar_types, IsolationLevel, PoolSettings};
+pub use values::{ConnectionUri, IsolationLevel, PoolSettings, ResolvedSecret};
+pub use version2::occurring_scalar_types;
