+++ crates/configuration/src/values/mod.rs
@@ -1,5 +1,11 @@
+mod comparison;
+mod isolation_level;
+mod pool_settings;
 mod secret;
 mod uri;
 
+pub use comparison::ComparisonOperatorMapping;
+pub use isolation_level::IsolationLevel;
+pub use pool_settings::PoolSettings;
 pub use secret::ResolvedSecret;
 pub use uri::ConnectionUri;
