+++ crates/configuration/src/version2.rs
@@ -12,9 +12,10 @@ use ndc_sdk::connector;
 
 use query_engine_metadata::metadata;
 
-use crate::values::{ConnectionUri, ResolvedSecret};
+use crate::values::{
+    ComparisonOperatorMapping, ConnectionUri, IsolationLevel, PoolSettings, ResolvedSecret,
+};
 use crate::version1;
-pub use version1::PoolSettings;
 
 const CONFIGURATION_QUERY: &str = include_str!("version2.sql");
 
@@ -25,9 +26,9 @@ const CONFIGURATION_QUERY: &str = include_str!("version2.sql");
 pub struct RawConfiguration {
     // Connection string for a Postgres-compatible database
     pub connection_uri: ConnectionUri,
-    #[serde(skip_serializing_if = "version1::PoolSettings::is_default")]
+    #[serde(skip_serializing_if = "PoolSettings::is_default")]
     #[serde(default)]
-    pub pool_settings: version1::PoolSettings,
+    pub pool_settings: PoolSettings,
     #[serde(default)]
     pub isolation_level: IsolationLevel,
     #[serde(default)]
@@ -40,7 +41,7 @@ impl RawConfiguration {
     pub fn empty() -> Self {
         Self {
             connection_uri: ConnectionUri::Uri(ResolvedSecret("".to_string())),
-            pool_settings: version1::PoolSettings::default(),
+            pool_settings: PoolSettings::default(),
             isolation_level: IsolationLevel::default(),
             metadata: metadata::Metadata::default(),
             configure_options: ConfigureOptions::default(),
@@ -60,20 +61,6 @@ pub fn default_unqualified_schemas_for_types_and_procedures() -> Vec<String> {
         "pg_catalog".to_string(),
         "tiger".to_string(),
     ]
-}
-
-/// The isolation level of the transaction in which a query is executed.
-#[derive(
-    Debug, Clone, Copy, Default, serde::Deserialize, serde::Serialize, schemars::JsonSchema,
-)]
-pub enum IsolationLevel {
-    /// Prevents reading data from another uncommitted transaction.
-    #[default]
-    ReadCommitted,
-    /// Reading the same data twice is guaranteed to return the same result.
-    RepeatableRead,
-    /// Concurrent transactions behave identically to serializing them one at a time.
-    Serializable,
 }
 
 /// Options which only influence how the configuration server updates the configuration
@@ -101,8 +88,8 @@ pub struct ConfigureOptions {
     #[serde(default = "default_unqualified_schemas_for_types_and_procedures")]
     pub unqualified_schemas_for_types_and_procedures: Vec<String>,
     /// The mapping of comparison operator names to apply when updating the configuration
-    #[serde(default = "version1::default_comparison_operator_mapping")]
-    pub comparison_operator_mapping: Vec<version1::ComparisonOperatorMapping>,
+    #[serde(default = "ComparisonOperatorMapping::default_mappings")]
+    pub comparison_operator_mapping: Vec<ComparisonOperatorMapping>,
     /// Which version of the generated mutation procedures to include in the schema response
     #[serde(default)]
     pub mutations_version: Option<metadata::mutations::MutationsVersion>,
@@ -124,7 +111,7 @@ impl Default for ConfigureOptions {
             unqualified_schemas_for_tables: default_unqualified_schemas_for_tables(),
             unqualified_schemas_for_types_and_procedures:
                 default_unqualified_schemas_for_types_and_procedures(),
-            comparison_operator_mapping: version1::default_comparison_operator_mapping(),
+            comparison_operator_mapping: ComparisonOperatorMapping::default_mappings(),
             // we'll change this to `Some(MutationsVersions::V1)` when we
             // want to "release" this behaviour
             mutations_version: None,
